requirements.txt
azure-functions
pandas
azure-cosmosdb-table
azure-keyvault-secrets
azure-identity